remove dangling code, and remove duplicate copy of PendingPromise
